Add files via upload
SOFTWARE DESIGN PROJECT:
The purpose of this lab test is to implement the TemplateMethod, the Singleton, the Factory and the Command patterns in one
application.
The GeneralTemplate objects (i.e. GeneralTemplate, TemplateA, TemplateB, TemplateC) play the
role of the “receiver” in the Command pattern.
The different GeneralTemplate objects are generated by a factory (use the repo method). The input
ID “A” by the reader is associated with a TemplateAFactory which generates a TemplateA class.
The similar idea is for input ID “B” and “C” (I.e. The input “B” by the reader is associated with a
TemplateBFactory which generates a TemplateB class etc.). Use a Factory Repo. See the
TemplateIDReader class for getting the IDs form the user. You can hard code the input returned to
save time. Run your program with all possible inputs i.e. “A”, “B”, and “C”.
Note that there is only one Invoker (i.e. TheInvoker) and only one Command (i.e. CommandA). 
Sequence of Actions: 
Client populates factory repo, gets input from user, creates invoker, creates Command, gets
appropriate factory and template, and links them so that they implement the Command design pattern. 

SAMPLE OUTPUT: 
If the input is “A” then the output is:
Starting Transaction in transaction class
Doing PartA in transaction class
Doing PartB in TemplateA class
Doing PartC in transaction class
If the input is “B” then the output is:
Starting Transaction in transaction class
Doing PartA in TemplateB class
Doing PartB in transaction class
Doing PartC in transaction class
If the input is “C” then the output is:
Starting Transaction in transaction class
Doing PartA in TemplateC class
Doing PartB in transaction class
Doing PartC in templateC class