Add files via upload
EECS 3311 - SOFTWARE DESIGN - JAVA
Implement a system that uses a combination of Adapter, State, Factory, and if you can, also Singleton.
More specifically, consider you have a client class Transaction which would like to invoke a backend service offered by classes BackEndService1, BackEndService2, BackEndService3, by passing a request of type RequestObject. The Transaction is referenced and called by the TransactionClient class which has the main method as well. 
The client class Transaction has access to an Adapter (can be the singleton), and the Adapter has attached to it a State which can be any of AdapterState1, AdapterState2, AdapterState3. Each adapter state can invoke one of back-end service, and more specifically, AdapterState1 can invoke the service do1() of BackEndService1, AdapterState2 can invoke the service do2() of BackEndService2, and AdapterState3 can invoke the service do3() of BackEndService3. The choice of what state is associated with the adapter is given by a user selection which comes as an integer input from the user provided by the StateReader utility. For example, a user input of 1 can generate AdapterState1, while a user input 2 can generate AdapterState2, and a user input of 3 can generate AdapterState3. The correct adapter states can be generated by a Factory, and the adapter states can be singletons. So, in other words each AdapterState “adapts” its corresponding back-end service.