Add files via upload
EECS 3311 - SOFTWARE DESIGN - JAVA 
Consider the problem of Lab 4 with the aircraft GPS receivers facilitating navigation. The description is given for your convenience in Appendix I.
In this Lab you will implement the Factory pattern in order to a) create the right command given the user’s input (e.g. navigation, fuel prediction etc.) and; b) create the right state of the aircraft given some input you will get by the user (should have been by a sensor) regarding the altitude, and the gear up or down status. 
For this lab check out the following parts of the code:
With respect to the command factory start at lines 22-35 on the FMCServer.java at the client package (Lab4 project). Make sure you create the right packages to put your command factory classes
With resect to the state factory, look at line 73 in the NavigationFacade.java in the frontEnd package.
Think your design carefully. If you need to add new classes e.g. to hold data please feel free to do so or re-factor your code feel free to do so.