Add files via upload
EECS 3311 - SOFTWARE DESIGN - JAVA 
Consider the problem of Lab 3 with the aircraft GPS receivers facilitating navigation. The description is given for your convenience in Appendix I.
In this Lab we will extend the problem by considering the State design pattern and the Command Design pattern.
STATE design pattern
More specifically, we would like the navigation to be different according to the altitude of the aircraft as follows:
If the altitude is less than 1,000 ft, and the gear is not down then notify only the elevator actuator only
If the altitude is between 1,000ft and 1,500 feet and the gear is not down notify only the elevator and the ailerons actuators 
If the altitude is less than 500 ft and the gear is down do not do any notifications

Decide in which class the State design pattern will be applied on
Implement the pattern. Hint: Here you have three states

COMMAND design pattern
Make the navigate() command of the autopilot to be a “command” using the Command design pattern.
For that you will need:
An Invoker (i.e. a class which will execute() the commands)
An ICommand interface and a concrete command class NavigateCommand implementing this interface
A Receiver (this will be the Autopilot for the NavigateCommand)
In this respect we may have the autopilot (as a receiver) to be able to implement many different commands. 
Decide how the pattern will look like in this problem
Implement the pattern. Hint: For the time being, instantiate the invoker and the command on the NavigationServer and have as the Receiver the Autopilot which will have the method navigate() as in the previous Lab. 


APPENDIX I 
ORIGINAL PROBLEM DESCRIPTION

Problem Description
Consider an aircraft which has three GPS receivers obtaining at real time the position of the aircraft in terms of latitude and longitude values. The autopilot software of the aircraft sends requests to obtain the coordinates from the three GPS receivers. There are two brands of GPS receivers (two Honeywell and one Rockwell). The current system employs a 2-3 voting strategy to determine the coordinate to use (i.e. if two of the GPS readings are within a tolerance limit, then any of the two readings is used, and when all three readings differ at a higher than the tolerance value then an error message is generated and the reading of the fist GPS is used).  Once the current coordinates are obtained the appropriate modules to control surfaces of the airplane are notified so that the airplane can be steered to the right direction to reach the next waypoint which is stored in the autopilot. 
Assumptions: 
Consider the longitude and latitude values to be numbers from 0-2
Consider that the messages to modules that move the control surfaces (i.e. the rudder, the ailerons, the elevator) steer the aircraft in a proper way

Task:
Design the system as a collection of classes. You may see the current solution as a guideline. Apply the following design patterns: Proxy, Observer, Façade, Strategy.

Process:
The Autopilot object sends a message (do this on a loop 100 times) to Proxy component (see Proxy pattern). 
The Proxy object calls the Navigation Façade
The Navigation Façade does the following sequence of steps (see Façade pattern)
3.a Send a request to GPSReceiver1 object to get the first reading
3.b Send a request to GPSReceiver2 object to get the second reading
3.c Send a request to GPSReceiver3 object to get the third reading
3.d Apply the current comparison strategy (2-3 voting, or first reading is always correct) to get   
       the resulting coordinates (see the Strategy pattern)
3.e Update the Autopilot with the new coordinates 
3.f Send the coordinates to the GPS Date Subject
The GPS Data Subject notifies the Actuators (see the Observer Pattern)
The Autopilot notifies the control surfaces with the offset of how much left, right, up or down the airplane is off from the desired path (see the Observer pattern)
Hints:
Among others use an AutoPilot class, a Proxy class, a GPSReceiver class, a ICompareCoords interface, two coordinate comparison strategy classes TwoThreeVoting, FirstIsBest which implement the ICompareCoords interface, an IControlActuator interface, and three classes RudderActuator, AileronActuator,  ElevatorActuator classes (issue to move the control surfaces based on the signals the autopilot gets and its target waypoint) which implement the IControlActuator interface.